[REFACTOR] capsule context and propvider to Separate file
